Fixed integer overflows flattening shapes (e.g., 32,32,2048,2048).

+++ h5py/_hl/attrs.py
@@ -156,7 +156,7 @@ class AttributeManager(base.MutableMappingHDF5, base.CommonStateObject):
             # is compatible, and reshape if needed.
             else:
 
-                if shape is not None and numpy.product(shape) != numpy.product(data.shape):
+                if shape is not None and numpy.product(shape, dtype=numpy.uint64) != numpy.product(data.shape, dtype=numpy.uint64):
                     raise ValueError("Shape of new attribute conflicts with shape of data")
 
                 if shape != data.shape:
@@ -229,7 +229,7 @@ class AttributeManager(base.MutableMappingHDF5, base.CommonStateObject):
 
                 # Allow the case of () <-> (1,)
                 if (value.shape != attr.shape) and not \
-                   (numpy.product(value.shape) == 1 and numpy.product(attr.shape) == 1):
+                   (numpy.product(value.shape, dtype=numpy.uint64) == 1 and numpy.product(attr.shape, dtype=numpy.uint64) == 1):
                     raise TypeError("Shape of data is incompatible with existing attribute")
                 attr.write(value)
